server/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
APScheduler==3.10.4
